+++ BambooTracker/gui/command/instrument/change_instrument_name_qt_command.cpp
@@ -28,14 +28,13 @@
 #include "gui/instrument_editor/instrument_editor.hpp"
 #include "instrument_command_qt_utils.hpp"
 
-ChangeInstrumentNameQtCommand::ChangeInstrumentNameQtCommand(
-		QListWidget* list, int num, int row, std::weak_ptr<InstrumentEditorManager> dialogMan,
+ChangeInstrumentNameQtCommand::ChangeInstrumentNameQtCommand(QListWidget* list, int num, int row, std::weak_ptr<InstrumentEditorManager> editorMan,
 		const QString& oldName, const QString& newName, QUndoCommand* parent)
 	: QUndoCommand(parent),
 	  list_(list),
 	  num_(num),
 	  row_(row),
-	  dialogMan_(dialogMan),
+	  editorMan_(editorMan),
 	  oldName_(oldName),
 	  newName_(newName)
 {
@@ -47,9 +46,7 @@ void ChangeInstrumentNameQtCommand::redo()
 	auto title = gui_command_utils::makeInstrumentListText(num_, newName_);
 	item->setText(title);
 
-	if (auto dialog = dialogMan_.lock()->getDialog(num_)) {
-		dialog->setWindowTitle(title);
-	}
+	editorMan_.lock()->onInstrumentNameChanged(num_);
 }
 
 void ChangeInstrumentNameQtCommand::undo()
@@ -58,9 +55,7 @@ void ChangeInstrumentNameQtCommand::undo()
 	auto title = gui_command_utils::makeInstrumentListText(num_, oldName_);
 	item->setText(title);
 
-	if (auto dialog = dialogMan_.lock()->getDialog(num_)) {
-		dialog->setWindowTitle(title);
-	}
+	editorMan_.lock()->onInstrumentNameChanged(num_);
 }
 
 int ChangeInstrumentNameQtCommand::id() const