+++ BambooTracker/gui/command/instrument/swap_instruments_qt_command.hpp
@@ -39,7 +39,7 @@ class SwapInstrumentsQtCommand final : public QUndoCommand
 public:
 	SwapInstrumentsQtCommand(QListWidget *list, int inst1Row, int inst2Row,
 							 const QString& inst1Name, const QString& inst2Name,
-							 std::weak_ptr<InstrumentEditorManager> dialogMan,
+							 std::weak_ptr<InstrumentEditorManager> editorMan,
 							 PatternEditor* pattern, QUndoCommand* parent = nullptr);
 	void undo() override;
 	void redo() override;
@@ -49,7 +49,7 @@ private:
 	QListWidget* list_;
 	PatternEditor* ptn_;
 	int inst1Row_, inst2Row_;
-	std::weak_ptr<InstrumentEditorManager> dialogMan_;
+	std::weak_ptr<InstrumentEditorManager> editorMan_;
 	QString inst1Name_, inst2Name_;
 
 	void swap(int above, int below, QString aboveName, QString belowName);