+++ BambooTracker/gui/instrument_editor/adpcm_instrument_editor.cpp
@@ -30,7 +30,6 @@
 #include "gui/event_guard.hpp"
 #include "gui/jam_layout.hpp"
 #include "gui/instrument_editor/instrument_editor_utils.hpp"
-#include "gui/gui_utils.hpp"
 
 AdpcmInstrumentEditor::AdpcmInstrumentEditor(int num, QWidget* parent)
 	: InstrumentEditor(num, parent),
@@ -408,8 +407,7 @@ void AdpcmInstrumentEditor::updateInstrumentParameters()
 {
 	Ui::EventGuard eg(isIgnoreEvent_);
 
-	auto name = gui_utils::utf8ToQString(bt_.lock()->getInstrument(instNum_)->getName());
-	setWindowTitle(QString("%1: %2").arg(instNum_, 2, 16, QChar('0')).toUpper().arg(name));
+	updateWindowTitle();
 
 	setInstrumentSampleParameters();
 	setInstrumentEnvelopeParameters();