+++ BambooTracker/gui/instrument_editor/fm_instrument_editor.cpp
@@ -783,8 +783,7 @@ void FmInstrumentEditor::updateInstrumentParameters()
 {
 	Ui::EventGuard eg(isIgnoreEvent_);
 
-	auto name = gui_utils::utf8ToQString(bt_.lock()->getInstrument(instNum_)->getName());
-	setWindowTitle(QString("%1: %2").arg(instNum_, 2, 16, QChar('0')).toUpper().arg(name));
+	updateWindowTitle();
 
 	setInstrumentEnvelopeParameters();
 	setInstrumentLFOParameters();